# Properly transform light rect and occluder rect to perform Light2D culling in canvas space