Fixes: https://github.com/godotengine/godot/issues/100664

There were two problems reported by #100664:
1. Weird crash when adding a light occluder
2. Shadows not appearing

The crash was caused by incrementing the occluder count after checking if the occluder had a valid polygon. Later, we copy over the transform for each occluder regardless of whether the polygon exists, so the occluder count needs to increment for each occluder, not each _valid_ occluder.

The shadows not appearing came from false negatives when culling occluders against individual lights. The culling was done using the rect of the light against the aabb (rect) of the occluder which are both in their own local spaces. If the item transforms were similar, then this didn't create any issues. But when the transforms differ enough, then the culling fails when it shouldn't. The solution was to transform both rects into canvas space and do culling in canvas space. For the `light->rect_cache` I was able to do the transformation early and save a bunch of transformations later (we were doing the transformation 2 times for each light + once for each light for each canvasitem on screen), now we do it only once per frame per light). 

This further optimizes the CPU cost of using PointLight2Ds. On my M2 MBP, this increases the FPS in a release build of my light performance benchmark from 550 FPS to 650 FPS and from 400 FPS to 500 FPS in dev builds. The benchmark is from https://github.com/godotengine/godot/pull/100302